+++ pdfalyzer/__init__.py
@@ -16,7 +16,6 @@ if not environ.get('INVOKED_BY_PYTEST', False):
 
 from rich.columns import Columns
 from rich.panel import Panel
-from yaralyzer.config import YaralyzerConfig
 from yaralyzer.helpers.rich_text_helper import prefix_with_plain_text_obj
 from yaralyzer.output.file_export import invoke_rich_export
 from yaralyzer.output.rich_console import console
@@ -26,7 +25,7 @@ from pdfalyzer.output.pdfalyzer_presenter import PdfalyzerPresenter
 from pdfalyzer.output.styles.rich_theme import PDFALYZER_THEME_DICT
 from pdfalyzer.pdfalyzer import Pdfalyzer
 from pdfalyzer.util.pdf_parser_manager import PdfParserManager
-from pdfalyzer.util.argument_parser import ALL_STREAMS, output_sections, parse_arguments
+from pdfalyzer.util.argument_parser import output_sections, parse_arguments
 
 # For the table shown by running pdfalyzer_show_color_theme
 MAX_THEME_COL_SIZE = 35
@@ -34,41 +33,21 @@ MAX_THEME_COL_SIZE = 35
 
 def pdfalyze():
     args = parse_arguments()
-    PdfalyzerConfig._parsed_args = args
     pdfalyzer = Pdfalyzer(args.file_to_scan_path)
     pdfalyzer = PdfalyzerPresenter(pdfalyzer)
     output_basepath = None
 
     # Binary stream extraction is a special case
     if args.extract_binary_streams:
-        log_and_print(f"Extracting all binary streams in '{args.file_to_scan_path}' to files in '{args.output_dir}'...")
+        log_and_print(f"Extracting binary streams in '{args.file_to_scan_path}' to files in '{args.output_dir}'...")
         PdfParserManager(args.file_to_scan_path).extract_all_streams(args.output_dir)
         log_and_print(f"Binary stream extraction complete, files written to '{args.output_dir}'.\nExiting.\n")
         sys.exit()
 
-    def get_output_basepath(export_method):
-        """Build the path to an output file - everything but the extension"""
-        export_type = export_method.__name__.removeprefix('print_')
-        output_basename = f"{args.output_basename}.{export_type}"
-
-        if export_type == 'font_info':
-            output_basename += '_'
-
-            if args.streams != ALL_STREAMS:
-                output_basename += f"_id{args.streams}"
-
-            output_basename += f"_maxdecode{YaralyzerConfig.MAX_DECODE_LENGTH}"
-
-            if args.extract_quoteds:
-                output_basename += f"_quote_{','.join(args.extract_quoteds)}"
-
-        output_basename += args.file_suffix
-        return path.join(args.output_dir, output_basename + f"___pdfalyzed_{args.invoked_at_str}")
-
     # Analysis exports wrap themselves around the methods that actually generate the analyses
     for (arg, method) in output_sections(args, pdfalyzer):
         if args.output_dir:
-            output_basepath = get_output_basepath(method)
+            output_basepath = PdfalyzerConfig.get_output_basepath(method)
             print(f'Exporting {arg} data to {output_basepath}...')
             console.record = True