+++ pdfalyzer/binary/binary_scanner.py
@@ -3,20 +3,14 @@ Class for handling binary data - scanning through it for various suspicious patt
 various character encodings upon it to see what comes out.
 """
 from collections import defaultdict
-from numbers import Number
-from typing import Any, Iterator, Optional, Pattern, Tuple, Union
+from typing import Iterator, Optional, Tuple
 
-from deprecated import deprecated
-from rich.markup import escape
 from rich.panel import Panel
-from rich.table import Table
 from rich.text import Text
 from yaralyzer.bytes_match import BytesMatch
-from yaralyzer.config import YaralyzerConfig
 from yaralyzer.decoding.bytes_decoder import BytesDecoder
 from yaralyzer.encoding_detection.character_encodings import BOMS
 from yaralyzer.helpers.bytes_helper import hex_string, print_bytes
-from yaralyzer.helpers.rich_text_helper import CENTER, na_txt, prefix_with_plain_text_obj
 from yaralyzer.helpers.string_helper import escape_yara_pattern
 from yaralyzer.output.rich_console import BYTES_NO_DIM, console, console_width
 from yaralyzer.output.regex_match_metrics import RegexMatchMetrics
@@ -26,36 +20,46 @@ from yaralyzer.util.logging import log
 
 from pdfalyzer.config import PdfalyzerConfig
 from pdfalyzer.decorators.pdf_tree_node import PdfTreeNode
-from pdfalyzer.detection.constants.binary_regexes import (BACKTICK, DANGEROUS_STRINGS, FRONTSLASH,
-     GUILLEMET, QUOTE_PATTERNS)
+from pdfalyzer.detection.constants.binary_regexes import (BACKTICK,
+     DANGEROUS_PDF_KEYS_TO_HUNT_ONLY_IN_FONTS, DANGEROUS_STRINGS, FRONTSLASH, GUILLEMET,
+     QUOTE_PATTERNS)
 from pdfalyzer.helpers.string_helper import generate_hyphen_line
-from pdfalyzer.output.layout import (generate_subtable, half_width, pad_header, print_headline_panel,
-     print_section_sub_subheader)
-from pdfalyzer.util.adobe_strings import CONTENTS, CURRENTFILE_EEXEC
+from pdfalyzer.output.layout import print_headline_panel, print_section_sub_subheader
+from pdfalyzer.util.adobe_strings import CONTENTS, CURRENTFILE_EEXEC, FONT_FILE_KEYS
 
 
 class BinaryScanner:
-    def __init__(self, _bytes: bytes, owner: Union['FontInfo', 'PdfTreeNode'], label: Optional[Text] = None):
+    def __init__(self, _bytes: bytes, owner: PdfTreeNode, label: Optional[Text] = None):
         """owner is an optional link back to the object containing this binary"""
         self.bytes = _bytes
         self.label = label
         self.owner = owner
+        self.stream_length = len(_bytes)
 
         if label is None and isinstance(owner, PdfTreeNode):
              self.label = owner.__rich__()
 
-        self.stream_length = len(_bytes)
-        self.regex_extraction_stats = defaultdict(lambda: RegexMatchMetrics())
         self.suppression_notice_queue = []
+        self.regex_extraction_stats = defaultdict(lambda: RegexMatchMetrics())
 
     def check_for_dangerous_instructions(self) -> None:
         """Scan for all the strings in DANGEROUS_INSTRUCTIONS list and decode bytes around them"""
-        print_section_sub_subheader("Scanning Binary For Anything 'Mad Sus'...", style=f"bright_red")
+        subheader = "Scanning Binary For Anything That Could Be Described As 'Sus'..."
+        print_section_sub_subheader(subheader, style=f"bright_red")
 
         for instruction in DANGEROUS_STRINGS:
-            yaralyzer = self._pattern_yaralyzer(instruction, REGEX)
+            yaralyzer = self._pattern_yaralyzer(instruction, REGEX)  # TODO maybe change REGEX const string?
             yaralyzer.highlight_style = 'bright_red bold'
             self.process_yara_matches(yaralyzer, instruction, force=True)
+
+        # TODO code smell: This check should probably be in the calling code not here in the instance method
+        if self.owner.type in FONT_FILE_KEYS:
+            log.info(f"{self.owner} is a /FontFile. Scanning for short but dangerous PDF keys...")
+
+            for instruction in DANGEROUS_PDF_KEYS_TO_HUNT_ONLY_IN_FONTS:
+                yaralyzer = self._pattern_yaralyzer(instruction, REGEX)
+                yaralyzer.highlight_style = 'bright_red bold'
+                self.process_yara_matches(yaralyzer, instruction, force=True)
 
     def check_for_boms(self) -> None:
         """Check the binary data for BOMs"""
@@ -71,10 +75,11 @@ class BinaryScanner:
         Find all strings matching QUOTE_PATTERNS (AKA between quote chars) and decode them with various encodings.
         The --quote-type arg will limit this decode to just one kind of quote.
         """
-        quote_selections = PdfalyzerConfig.args().extract_quoteds
+        quote_selections = PdfalyzerConfig._args.extract_quoteds
 
         if len(quote_selections) == 0:
-            console.print("\nNo --extract-quoted options provided. Skipping extract/decode of quoted bytes.")
+            headline = "Skipping extract/decode of quoted bytes (--extract-quoted is empty)"
+            print_section_sub_subheader(headline, style='grey')
 
         for quote_type in quote_selections:
             if self.owner and self.owner.type == CONTENTS and quote_type in [FRONTSLASH, GUILLEMET]:
@@ -104,8 +109,8 @@ class BinaryScanner:
         return self._quote_yaralyzer(QUOTE_PATTERNS[FRONTSLASH], FRONTSLASH).match_iterator()
 
     def print_stream_preview(self, num_bytes=None, title_suffix=None) -> None:
-        """Print a preview showing the beginning and end of the stream data"""
-        num_bytes = num_bytes or console_width()
+        """Print a preview showing the beginning and end of the embedded stream data"""
+        num_bytes = num_bytes or PdfalyzerConfig._args.preview_stream_length or console_width()
         snipped_byte_count = self.stream_length - (num_bytes * 2)
         console.line()
 
@@ -125,26 +130,25 @@ class BinaryScanner:
             console.print(f"\n    <...skip {snipped_byte_count} bytes...>\n", style='dim')
             print_bytes(self.bytes[-num_bytes:])
 
-        console.print(generate_hyphen_line(title=title), style='dim')
+        console.print(generate_hyphen_line(title="END " + title), style='dim')
         console.line()
 
     def process_yara_matches(self, yaralyzer: Yaralyzer, pattern: str, force: bool = False) -> None:
         """Decide whether to attempt to decode the matched bytes, track stats. force param ignores min/max length"""
-        for bytes_match, bytes_decoder in yaralyzer.match_iterator():
-            log.debug(f"Trackings stats for match: {pattern}, bytes_match: {bytes_match}")
-            self.regex_extraction_stats[pattern].match_count += 1
-            self.regex_extraction_stats[pattern].bytes_matched += bytes_match.match_length
-            self.regex_extraction_stats[pattern].bytes_match_objs.append(bytes_match)
+        for bytes_match, decoder in yaralyzer.match_iterator():
+            log.debug(f"Trackings stats for match: {pattern}, bytes_match: {bytes_match}, is_decodable: {bytes_match.is_decodable()}")
 
             # Send suppressed decodes to a queue and track the reason for the suppression in the stats
-            if not ((YaralyzerConfig.MIN_DECODE_LENGTH < bytes_match.match_length < YaralyzerConfig.MAX_DECODE_LENGTH) \
-                    or force):
-                self._queue_suppression_notice(bytes_match, pattern)
+            if not (bytes_match.is_decodable() or force):
+                self.suppression_notice_queue.append(bytes_match.suppression_notice())
                 continue
 
             # Print out any queued suppressed notices before printing non suppressed matches
             self._print_suppression_notices()
-            self._record_decode_stats(bytes_match, bytes_decoder, pattern)
+            console.print(decoder)
+            self.regex_extraction_stats[pattern].tally_match(decoder) # TODO: This call must come after print(decoder)
+
+        self._print_suppression_notices()
 
         # This check initializes the defaultdict for 'pattern'
         if self.regex_extraction_stats[pattern].match_count == 0:
@@ -179,38 +183,6 @@ class BinaryScanner:
             rules_label=safe_label(rules_label or pattern),  # TODO: maybe safe_label() belongs in yaralyzer
             pattern_label=safe_label(pattern_label or pattern)
         )
-
-    def _record_decode_stats(self, bytes_match: BytesMatch, decoder: BytesDecoder, pattern: str) -> None:
-        """Attempt to decode _bytes with all configured encodings and print a table of the results"""
-        # Track stats on whether the bytes were decodable or not w/a given encoding
-        log.debug(f"  Recording stats for {bytes_match} (pattern: {pattern})")
-        self.regex_extraction_stats[pattern].matches_decoded += 1
-
-        for encoding, count in decoder.was_match_decodable.items():
-            decode_stats = self.regex_extraction_stats[pattern].was_match_decodable
-            decode_stats[encoding] = decode_stats.get(encoding, 0) + count
-
-        for encoding, count in decoder.was_match_undecodable.items():
-            failure_stats = self.regex_extraction_stats[pattern].was_match_undecodable
-            failure_stats[encoding] = failure_stats.get(encoding, 0) + count
-
-        for encoding, count in decoder.was_match_force_decoded.items():
-            forced_stats = self.regex_extraction_stats[pattern].was_match_force_decoded
-            forced_stats[encoding] = forced_stats.get(encoding, 0) + count
-
-    def _queue_suppression_notice(self, bytes_match: BytesMatch, quote_type: str) -> None:
-        """Print a message indicating that we are not going to decode a given block of bytes"""
-        self.regex_extraction_stats[bytes_match.label].skipped_matches_lengths[bytes_match.match_length] += 1
-        txt = bytes_match.__rich__()
-
-        if bytes_match.match_length < YaralyzerConfig.MIN_DECODE_LENGTH:
-            txt = Text('Too little to actually attempt decode at ', style='grey') + txt
-        else:
-            txt.append(" too long to decode ")
-            txt.append(f"(--max-decode-length is {YaralyzerConfig.MAX_DECODE_LENGTH} bytes)", style='grey')
-
-        log.debug(Text('Queueing suppression notice: ') + txt)
-        self.suppression_notice_queue.append(txt)
 
     def _print_suppression_notices(self) -> None:
         """Print notices in queue in a single panel; empty queue"""