+++ tests/test_file_export.py
@@ -22,7 +22,7 @@ def test_file_export(analyzing_malicious_pdf_path, tmp_dir):
     rendered_files = files_in_dir(tmp_dir)
     assert len(rendered_files) == 6
     file_sizes = sorted([path.getsize(f) for f in rendered_files])
-    assert_array_is_close(file_sizes, [6905, 8356, 35908, 104811, 181688, 1689983])
+    assert_array_is_close(file_sizes, [6905, 8356, 35908, 170612, 181688, 1689983])
 
     for file in rendered_files:
         remove(file)